Extract copy/paste styles
